==> RAnalysis/Scripts/Popgen/08_kinship_matrices.Rmd <==
---
title: "07_kinship_matrices"
author: "Samuel Gurr"
---

## Objective: 

* ultimately the goal is to determine loci under selection and their heritability across generations - one method being QTL (quantitative trait locus) using packages such as lme4qtl. To run this analysis (scripts downstream from here) we need to assemble a covariance matrix based on the pedigree of our scallops. **In this script** will load the pedigree information to assemble kinship matrices 


* kinship matrix
  * use the outputs from ```sequoia``` to build kinship matrices necessary to run heritability and selection
  * **Important!**: kinship2 is a common R package to execuate this task, however is modelled on sexually dimporphic pedigree, meaning sires and dams are always unique individuals (written for humans). Our data on the bay scallop has some cases at which eggs and sperm were used from the same individual and alowed to batch fertilize separate to avoid selfing - in summary, we have ids that may have been identified as *both* dam and sire for different individuals and infact may be true (based on real batch fertilizations). Therefore we need to explore alternative tools other than kinship2
  * R packages to run kinship matrix: 
    - kinship2: as explained above, we will get errors, important to run anyway! Jesca wrote the function sequoia::CalcRPed that acts as a wrapper of kinship2::kinship, lets run this
    - ribd: seems like this is the option we should puruse, try it! 

### set working directory

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# SET WORKING DIRECTORY 
knitr::opts_knit$set(root.dir = "C:/Users/samjg/Documents/Github_repositories/EAD-ASEB-Airradians_Popgen_OA/") # Sam's
#knitr::opts_knit$set(root.dir = "C:/Users/samuel.gurr/Documents/Github_repositories/EAD-ASEB-Airradians_Popgen_OA/") # Sam's
```

### load libraries

```{r, include=FALSE}
library(tidyverse)
library(kinship2)
library(ribd) #install.packages('ribd')
library(sequoia)
```

### load pedigree output from  ```sequoia``` 

* load the pedigree

```{r load output pedigrees}

# pedigree :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

# F1 Offspring (includes low and moderate with F0 dams and sires, no F0 pedigree known)
F1O.pedigree <- read.csv("RAnalysis/Output/Popgen/pedigree/F0BF1O/F0BF1O_final.csv") %>% 
                      select(id, dam, sire) 

# F2 offspring (includes the juveniles and broodstock F2s, no F1 broodstock pedigree included)
F1BF2O.pedigree.low  <- read.csv("RAnalysis/Output/Popgen/pedigree/F1BF2O/Low/F1BF2O_Low_final.csv") %>% 
                                select(id, dam, sire) 
F1BF2O.pedigree.mod  <- read.csv("RAnalysis/Output/Popgen/pedigree/F1BF2O/Moderate/F1BF2O_Mod_final.csv") %>% 
                                select(id, dam, sire) 
F1BF2O.pedigree.high <- read.csv("RAnalysis/Output/Popgen/pedigree/F1BF2O/High/F1BF2O_High_final.csv") %>% 
                                select(id, dam, sire) 

# have not run these yet!
# F2BF3O.pedigree.low  <- read.csv("RAnalysis/Output/Popgen/pedigree/F2BF3O/Low/F2BF3O_Low_final.csv") %>% 
#                                 select(id, dam, sire) 
# F2BF3O.pedigree.mod  <- read.csv("RAnalysis/Output/Popgen/pedigree/F2BF3O/Moderate/F2BF3O_Mod_final.csv") %>% 
#                                 select(id, dam, sire) 
# F2BF3O.pedigree.high <- read.csv("RAnalysis/Output/Popgen/pedigree/F2BF3O/High/F2BF3O_High_final.csv") %>% 
#                                 select(id, dam, sire) 

# F2 offspring (includes the juveniles and broodstock F2s, no F1 broodstock pedigree included)


# F1, F2, and F3 low v high pedigree, note no F0 pedigree known or included in here
LowvMod.pedigree <- read.csv("RAnalysis/Output/Popgen/pedigree/All_Low_v_Moderate_final.csv") %>% 
                                select(id, dam, sire) 

```

### load life history data and subset as appropriate for pedigrees


* load the vcf file to otbain a list of all IDs
* assign birthyear and yearlast based on generation F0, F1, F2, and F3 in the id names
* load the parentage data of the F1 F2 and F3 offspring, these contain the broodstock "assignments" 
  * **note** assignments as female if only egg and male if only sperm was used fir batch fertilizations, otherwise given unknown for hermaphroditic contributions to batch fertilization, this may be inappropriate for these data and solely assign all animals as unknown for consistency - repeat this to test whether results change 
* subset All.LifeHistData.2 based on ids in the pedigree

```{r list all IDs in the master vcf}

# call th vcf file with all ids for the project 
All.vcf <- vcfR::read.vcfR(here::here(getwd(),
                            "RAnalysis", 
                            "Data", 
                            "Popgen", 
                            "03_prune", 
                            "out.7.phased.vcf.gz"), verbose = FALSE)
All.ListIDs           <- colnames(All.vcf@gt[,2:length(colnames(All.vcf@gt))])

```


```{r lifehistory and sex}

# use ids to creat ethe lifehostyr reference data

# data.frame with up to 6 columns:

# ID max. 30 characters long

# Sex 1 = female, 2 = male, 3 = unknown, 4 = hermaphrodite, other numbers or NA = unknown

# BirthYear birth or hatching year, integer, with missing values as NA or any negative number.

# BY.min minimum birth year, only used if BirthYear is missing
 
# BY.max maximum birth year, only used if BirthYear is missing
 
# Year.last  Last year in which individual could have had offspring. Can e.g. in mammals be the year before death for females, and year after death for males.


# OBJECTIVE:
# * below we (1) manually create a lifehist data frame assigning sex as '4' hermaphordites
# (2) merge all parentage broodstock data where we known whether eggs, sperm, or both were used and thus 1,2 or 4 for sex assignment
# (3) lastly merge the first manual lifehist dataframe with the adult broodstock sex assignment for a final 'editted' data

# 1. LIFE HISTPRY DATA  :::::::::::::::::::::::::::
# * make this manually based on numeric birthyear
# * here we assume the F0s start as year 1 and are 1 year olds, meaning they could spawn until year 3,
# * the remaining generation s follow the same pattern, with viable spawning history fro two years after birth
patterns <- c("F0", "F1", "F2","F3") # a few IDs do not have a 'F' delimiter for generation, these are all F1 juveniles
All.LifeHistData.1 <- as.data.frame(matrix(NA, 
                                             ncol = 6, 
                                             nrow = length(All.ListIDs))) %>% 
                          dplyr::rename(id=V1,
                                 sex=V2, # make this 4, hermaphordite
                                 BirthYear=V3,
                                 BY.min=V4,# only need if BirthYear is missing, NA
                                 BY.max=V5,# only need if BirthYear is missing, NA
                                 YearLast=V6) %>% 
                          dplyr::mutate(id = All.ListIDs,
                                 sex = 4,
                                 BirthYear = case_when(
                                   grepl("F0", id) ~ 1,
                                   !grepl(paste(patterns,collapse="|"), id) ~ 2, # a few F1 juveniles without 'F' in ID, birthyear 2
                                   grepl("F1", id) ~ 2,
                                   grepl("F2", id) ~ 3,
                                   grepl("F3", id) ~ 4),
                                   
                                 YearLast = case_when(
                                   grepl("F0", id) ~ 3,
                                   !grepl(paste(patterns,collapse="|"), id) ~ 4, # a few F1 juveniles without 'F' in ID, birthyear 2
                                   grepl("F1", id) ~ 4,
                                   grepl("F2", id) ~ 5,
                                   grepl("F3", id) ~ 6),
                          )


# 2. SEX  :::::::::::::::::::::::::::
# * note that we have controlled spawns where we used eggs and or sperm of the spawners and recorded this information 
# * although scallops are hermaphrodites (noted as '4' for all offspring!) we can mark the broodstock as dams sires or both

F1.Parentage <- read.csv(paste0(getwd(),"/RAnalysis/Data/Popgen/Parentages/F1_parentage.csv"), sep = ',', head = T) %>% 
                  dplyr::mutate(Gen = "F0",
                                Treatment = NA) # added just to match F2 and F3.Parentage and rbind a master below

F2.Parentage <- read.csv(paste0(getwd(),"/RAnalysis/Data/Popgen/Parentages/F2_parentage.csv"), sep = ',', head = T) %>% 
                  dplyr::mutate(Gen = "F1",
                                Batch_5 = NA) # added just to match F1.Parentage and rbind a master below

F3.Parentage <- read.csv(paste0(getwd(),"/RAnalysis/Data/Popgen/Parentages/F3_parentage.csv"), sep = ',', head = T) %>% 
                  dplyr::mutate(Gen = "F2",
                                Batch_5 = NA) # added just to match F1.Parentage and rbind a master below

# first, run a sanity check before binding all the parentage datafiles
nrow( 
  rbind(
        (F1.Parentage %>%  select(vcf_ID_raw, Ped)),
        (F2.Parentage %>%  select(vcf_ID_raw, Ped)),
        (F3.Parentage %>%  select(vcf_ID_raw, Ped))
        )
    ) # 107 rows bind ing these data
nrow(F1.Parentage) + nrow(F2.Parentage) + nrow(F3.Parentage) # 107
# great, move forwad with binding and editting for life history data
# find the parentage data files and edit them 
All.Sex <-   rbind(
                  (F1.Parentage %>%  select(vcf_ID_raw, Ped)),
                  (F2.Parentage %>%  select(vcf_ID_raw, Ped)),
                  (F3.Parentage %>%  select(vcf_ID_raw, Ped))
                  ) %>% 
             dplyr::mutate(sex = case_when(
                              Ped %in% "Dam" ~ 1, # female = only eggs contribute to spawns
                              Ped %in% "Sire" ~ 2, # male  = only sperm contribute to spawns
                              Ped %in% "Both" ~ 4 # both eggs and sperm were used 
                            )) %>% 
             dplyr::select(!Ped) %>% 
             dplyr::rename(id = vcf_ID_raw)


# 3. MERGE LIFEHIST DATAFRAME WITH BROODSTOCK SEX ASSIGNMENT  :::::::::::::::::::::::::::
# now edit the lifehistdata including information for the Sex dataframe  
All.LifeHistData.2   <-  merge( 
                               dplyr::filter(All.LifeHistData.1, !grepl(c("-B"), id)), # absent of all broodstock
                               All.Sex, # merge in the broodstock data
                               all = T) %>% 
                            dplyr::mutate(BirthYear = case_when(
                                                                 grepl("F0", id) ~ 1,
                                                                 # a few F1 juveniles without 'F' in ID, birthyear 2
                                                                 !grepl(paste(patterns,collapse="|"), id) ~ 2, 
                                                                 grepl("F1", id) ~ 2,
                                                                 grepl("F2", id) ~ 3,
                                                                 grepl("F3", id) ~ 4),
                                   
                                          YearLast = case_when(
                                                                 grepl("F0", id) ~ 2,
                                                                 # a few F1 juveniles without 'F' in ID, birthyear 2
                                                                 !grepl(paste(patterns,collapse="|"), id) ~ 3, 
                                                                 grepl("F1", id) ~ 3,
                                                                 grepl("F2", id) ~ 4,
                                                                 grepl("F3", id) ~ 5)
                                 )
```


### load parentage data

* F1.Parentage - contains the parentage of the F1 offpsirng, meaning the female (egg) and male (sperm) ids and their batch fertilization ids of the  **F0 Broodstock**
  * *no pCO2 treamtents, 'founder' population of the experiment, F1 progeny (as embryos) were dispersed in low and moderate pCO2 treatments*

* F2.Parentage - contains the parentage of the F1 offpsirng, meaning the female (egg) and male (sperm) ids and their batch fertilization ids of the  **F1 Broodstock**
  * *low and moderate pCO2 treatments, contained matched progeny exposure*
  
* F3.Parentage - contains the parentage of the F1 offpsirng, meaning the female (egg) and male (sperm) ids and their batch fertilization ids of the  **F2 Broodstock**
  * *low,  moderate, and high (!) pCO2 treatments, contained matched progeny exposure*


* '_ped' = prepared the parentage data to rbind to the pedigree of their offpsring, formatted for kingship2::kinship()
  - format for kinship() must have four columns id, dadid, momid, sex
  - assign dam and sire (momid and dadid) as NA for the F0 broodstock
  - sex: bay scallops are simultaneous hermaphrodites, however kingship() faces errors when parents 
  have conflicting results with the dadid and momid assignment to progeny
```{r load broodstock pedigree for F0 data}
F1.Parentage <- read.csv(
                        paste0(
                        getwd(),
                        "/RAnalysis/Data/Popgen/Parentages/F1_parentage.csv"), 
                        sep = ',', 
                        head = T) 

F0B.pedigree <- F1.Parentage %>% 
                        dplyr::select(vcf_ID_raw,Ped) %>% 
                        dplyr::rename(id = vcf_ID_raw) %>% 
                        dplyr::mutate(sire = NA,
                                      dam = NA) %>% 
                                      # sex = case_when(Ped %in% 'Both' ~ 3,
                                      #                 Ped %in% 'Dam' ~ 1,
                                      #                 Ped %in% 'Sire' ~ 2)) %>% 
                        dplyr::select(!Ped)
```

### pedigree and kinship matrix: attempt sequoia::CalcRPed wrapper of kinship2

* **Note** as described at the top of the script, this will throw an error if there is an pedigree as dam and sire for different individuals (does not like hermaphrodites!)

```{r kinship2 will not work}
CalcRped(F0BF1O.pedigree, OUT = "DF") # Error message: Id not male, but is a father
CalcRped(F1BF2O.pedigree.low, OUT = "DF") # Error message: Id not male, but is a father adapter_trim.F1-B8-pH8.bam
CalcRped(F1BF2O.pedigree.mod, OUT = "DF") # ran successfully because there was not a case of hermaphrodite pedigree
CalcRped(F1BF2O.pedigree.high, OUT = "DF") # Id not male, but is a father: adapter_trim.F1-B8-pH8.bam
```

### pedigree and kinship matrix: use pedtools and ribd 

```{r about ribd}
?ribd
# Founders are allowed to be inbred, which enables construction of any given kappa coefficients, as described in Vigeland (2020). In addition to the standard coefficients, 'ribd' also computes a range of lesser-known coefficients, including generalised kinship coefficients, multi-person coefficients and two-locus coefficients
```

* reformat the pedigree
```{r F0 broodstock and F1 offspring pedtools::ped}
# bind the F0 pedigree (all unknown) with the known F1 offpsring pedigree (from sequoia)
# and merge with the lifehistory data to obtain sex
# sex == 4 was assigned for all offspring, whereas 1 or 2 was assigned for parents that solely contributed sperm and egg
# pedtools::ped only recognizes sex 0,1, and 2 where 0 is unknown
F0BF1O.pedigree <- merge(
                          rbind(F1O.pedigree, F0B.pedigree), # bind F1 offspring and F0 broodstokc pedigree
                          (All.LifeHistData.2 %>% select(id, sex)), # merge with lifehistory to get sex
                        by = 'id'
                        ) %>% 
                    #dplyr::mutate(sex == 0) # try this as well, just assign all as unknown since they are hermaphrodites
                    dplyr::mutate(sex =  # format sex so that all occurances of '4' are ow 0 for pedtools
                                    case_when(sex == 4 ~ 0,
                                              sex %in% NA ~ 0,
                                              .default = sex))

# ped
F0BF1O.pedform <- pedtools::ped(id = F0BF1O.pedigree$id,
                                        fid = F0BF1O.pedigree$dam,
                                        mid = F0BF1O.pedigree$sire,
                                        sex = F0BF1O.pedigree$sex)
nrow(as.data.frame(F0BF1O.pedform$`_comp1`))

# wrtie ped data and kinship
write.csv(as.matrix(F0BF1O.pedform$`_comp1`),
          paste0(getwd(),"/RAnalysis/Output/Popgen/pedigree/F0BF1O/F0BF1O_ped.csv"),
          row.names=FALSE)
```

```{r F1 broodstock and F2 offspring from LOW OA pedtools::ped}
# bind the F0 pedigree (all unknown) with the known F1 offpsring pedigree (from sequoia)
# and merge with the lifehistory data to obtain sex
# sex == 4 was assigned for all offspring, whereas 1 or 2 was assigned for parents that solely contributed sperm and egg
# pedtools::ped only recognizes sex 0,1, and 2 where 0 is unknown
F1BF2O.pedigree.sex <- merge(
                          F1BF2O.pedigree.low, # bind F1 offspring and F0 broodstokc pedigree
                          (All.LifeHistData.2 %>% select(id, sex)), # merge with lifehistory to get sex
                        by = 'id'
                        ) %>% 
                    #dplyr::mutate(sex == 0) # try this as well, just assign all as unknown since they are hermaphrodites
                    dplyr::mutate(sex =  # format sex so that all occurances of '4' are ow 0 for pedtools
                                    case_when(sex == 4 ~ 0,
                                              .default = sex))

# ped
F1BF2O_low.pedform <- pedtools::ped(id = F1BF2O.pedigree.sex$id,
                                        fid = F1BF2O.pedigree.sex$dam,
                                        mid = F1BF2O.pedigree.sex$sire,
                                        sex = F1BF2O.pedigree.sex$sex)

# wrtie ped data and kinship
write.csv(F0BF1O.pedform,
          paste0(getwd(),"/RAnalysis/Output/Popgen/pedigree/F0BF1O/F0BF1O_ped.csv"),
          row.names=FALSE)
```


```{r all offspring from low and moderate OA pedtools::ped}
# bind the F0 pedigree (all unknown) with the known F1 offpsring pedigree (from sequoia)
# and merge with the lifehistory data to obtain sex
# sex == 4 was assigned for all offspring, whereas 1 or 2 was assigned for parents that solely contributed sperm and egg
# pedtools::ped only recognizes sex 0,1, and 2 where 0 is unknown

# LowvMod.pedigree <- LowvMod.pedigree[-13,] # we have a duplicate F1-B25 id, need to address this in the pedigree script
LowvModerate.pedigree.sex <- merge(
                          rbind(LowvMod.pedigree, F0B.pedigree), # bind F1 offspring and F0 broodstokc pedigree
                          (All.LifeHistData.2 %>% select(id, sex)), # merge with lifehistory to get sex
                        by = 'id'
                        ) %>% 
                    #dplyr::mutate(sex == 0) # try this as well, just assign all as unknown since they are hermaphrodites
                    dplyr::mutate(sex =  # format sex so that all occurances of '4' are ow 0 for pedtools
                                    case_when(sex == 4 ~ 0,
                                              .default = sex))

View(LowvMod.pedigree)
# ped
LowvModerate.pedform <- pedtools::ped(id = LowvModerate.pedigree.sex$id,
                                        fid = LowvModerate.pedigree.sex$dam,
                                        mid = LowvModerate.pedigree.sex$sire,
                                        sex = LowvModerate.pedigree.sex$sex)

# wrtie ped data and kinship
write.csv(F0BF1O.pedform,
          paste0(getwd(),"/RAnalysis/Output/Popgen/pedigree/F0BF1O/F0BF1O_ped.csv"),
          row.names=FALSE)
```

* assemble a kinship matrix

```{r F0 broodstock F1 offspring pedtools::kinship}
F0BF1O.kinship <- kinship(F0BF1O.pedform)

# wrtie the kinship matrix
write.csv(as.matrix(F0BF1O.kinship),
          paste0(getwd(),"/RAnalysis/Output/Popgen/pedigree/F0BF1O/F0BF1O_kinship.csv"),
          row.names=TRUE)
```
